feat(lib): convert fee amount question to typescript
